+++ Engine/Core/ImGui/ImGui.cpp
@@ -7,6 +7,7 @@
 #include "ImUtil.h"
 #include "ImGuiAabb.h"
 #include "Core/Resource/Resource.h"
+#include "ImGuiInternal.h"
 
 // @formatter:off
 
@@ -38,55 +39,6 @@ static ImGuiWindow* FindHoveredWindow(ImVec2 pos, bool excluding_childs);
 
 }; // namespace ImGui
 
-struct ImGuiColMod	// Color/style modifier, backup of modified data so we can restore it
-{
-    ImGuiCol	Col;
-    ImVec4		PreviousValue;
-};
-
-// Temporary per-window data, reset at the beginning of the frame
-struct ImGuiDrawContext
-{
-    ImVec2					CursorPos{};
-    ImVec2					CursorPosPrevLine{};
-    ImVec2					CursorStartPos{};
-    float					CurrentLineHeight;
-    float					PrevLineHeight;
-    float					LogLineHeight;
-    int						TreeDepth;
-    bool					LastItemHovered;
-    ImVector<ImGuiWindow*>	ChildWindows;
-    ImVector<bool>			AllowKeyboardFocus;
-    ImVector<float>			ItemWidth;
-    ImVector<ImGuiColMod>	ColorModifiers;
-    ImGuiColorEditMode		ColorEditMode{};
-    ImGuiStorage*			StateStorage;
-    int						OpenNextNode;
-
-    float					ColumnStartX;
-    int						ColumnCurrent;
-    int						ColumnsCount;
-    bool					ColumnsShowBorders;
-    ImVec2					ColumnsStartCursorPos{};
-    ImGuiID					ColumnsSetID{};
-
-    ImGuiDrawContext()
-    {
-        CursorPos = CursorPosPrevLine = CursorStartPos = ImVec2(0.0f, 0.0f);
-        CurrentLineHeight = PrevLineHeight = 0.0f;
-        LogLineHeight = -1.0f;
-        TreeDepth = 0;
-        LastItemHovered = false;
-        StateStorage = NULL;
-        OpenNextNode = -1;
-
-        ColumnStartX = 0.0f;
-        ColumnCurrent = 0;
-        ColumnsCount = 1;
-        ColumnsShowBorders = true;
-        ColumnsStartCursorPos = ImVec2(0,0);
-    }
-};
 
 struct ImGuiTextEditState;
 #define STB_TEXTEDIT_STRING	ImGuiTextEditState
@@ -123,228 +75,6 @@ struct ImGuiTextEditState
     static void			RenderTextScrolledClipped(ImFont font, float font_size, const char* buf, ImVec2 pos_base, float width, float scroll_x);
 };
 
-struct ImGuiIniData
-{
-    char*	Name{};
-    ImVec2	Pos{};
-    ImVec2	Size{};
-    bool	Collapsed{};
-
-    ImGuiIniData() { memset(this, 0, sizeof(*this)); }
-    ~ImGuiIniData() { if (Name) { free(Name); Name = NULL; } }
-};
-
-struct ImGuiState
-{
-    bool					Initialized;
-    ImGuiIO					IO;
-    ImGuiStyle				Style;
-    float					Time;
-    int						FrameCount;
-    int						FrameCountRendered;
-    ImVector<ImGuiWindow*>	Windows;
-    ImGuiWindow*			CurrentWindow;						// Being drawn into
-    ImVector<ImGuiWindow*>	CurrentWindowStack;
-    ImGuiWindow*			FocusedWindow;						// Will catch keyboard inputs
-    ImGuiWindow*			HoveredWindow;						// Will catch mouse inputs
-    ImGuiWindow*			HoveredWindowExcludingChilds;		// Will catch mouse inputs (for focus/move only)
-    ImGuiID					HoveredId{};
-    ImGuiID					ActiveId{};
-    ImGuiID					ActiveIdPreviousFrame{};
-    bool					ActiveIdIsAlive;
-    float					SettingsDirtyTimer;
-    ImVector<ImGuiIniData*>	Settings;
-    ImVec2					NewWindowDefaultPos{};
-
-    // Render
-    ImVector<ImDrawList*>	RenderDrawLists;
-
-    // Widget state
-    ImGuiTextEditState		InputTextState;
-    ImGuiID					SliderAsInputTextId;
-    ImGuiStorage			ColorEditModeStorage;				// for user selection
-    ImGuiID					ActiveComboID;
-    char					Tooltip[1024]{};
-
-    // Logging
-    bool					LogEnabled;
-    FILE*					LogFile;
-    ImGuiTextBuffer			LogClipboard;
-    int						LogAutoExpandMaxDepth;
-
-    ImGuiState()
-    {
-        Initialized = false;
-        Time = 0.0f;
-        FrameCount = 0;
-        FrameCountRendered = -1;
-        CurrentWindow = NULL;
-        FocusedWindow = NULL;
-        HoveredWindow = NULL;
-        HoveredWindowExcludingChilds = NULL;
-        ActiveIdIsAlive = false;
-        SettingsDirtyTimer = 0.0f;
-        NewWindowDefaultPos = ImVec2(60, 60);
-        SliderAsInputTextId = 0;
-        ActiveComboID = 0;
-        memset(Tooltip, 0, sizeof(Tooltip));
-        LogEnabled = false;
-        LogFile = NULL;
-        LogAutoExpandMaxDepth = 2;
-    }
-};
-
-static ImGuiState	GImGui;
-
-struct ImGuiWindow
-{
-    char*					Name;
-    ImGuiID					ID;
-    ImGuiWindowFlags		Flags{};
-    ImVec2					PosFloat{};
-    ImVec2					Pos{};								// Position rounded-up to nearest pixel
-    ImVec2					Size{};								// Current size (==SizeFull or collapsed title bar size)
-    ImVec2					SizeFull{};							// Size when non collapsed
-    ImVec2					SizeContentsFit{};					// Size of contents (extents reach by the drawing cursor) - may not fit within Size.
-    float					ScrollY;
-    float					NextScrollY;
-    bool					ScrollbarY;
-    bool					Visible;
-    bool					Collapsed;
-    bool					Accessed{};
-    int						AutoFitFrames;
-
-    ImGuiDrawContext		DC;
-    ImVector<ImGuiID>		IDStack;
-    ImVector<ImVec4>		ClipRectStack;
-    int						LastFrameDrawn;
-    float					ItemWidthDefault;
-    ImGuiStorage			StateStorage;
-    float					FontScale;
-
-    int						FocusIdxCounter;					// Start at -1 and increase as assigned via FocusItemRegister()
-    int						FocusIdxRequestCurrent;				// Item being requested for focus, rely on layout to be stable between the frame pressing TAB and the next frame
-    int						FocusIdxRequestNext;				// Item being requested for focus, for next update
-
-    ImDrawList*				DrawList;
-
-public:
-    ImGuiWindow(const char* name, ImVec2 default_pos, ImVec2 default_size);
-    ~ImGuiWindow();
-
-    ImGuiID		GetID(const char* str);
-    ImGuiID		GetID(const void* ptr);
-
-    void		AddToRenderList();
-    bool		FocusItemRegister(bool is_active, int* out_idx = NULL);	// Return TRUE if focus is requested
-    void		FocusItemUnregister();
-
-    ImGuiAabb	Aabb() const							{ return ImGuiAabb(Pos, Pos+Size); }
-    ImFont		Font() const							{ return GImGui.IO.Font; }
-    float		FontSize() const						{ return GImGui.IO.FontHeight * FontScale; }
-    ImVec2		CursorPos() const						{ return DC.CursorPos; }
-    float		TitleBarHeight() const					{ return (Flags & ImGuiWindowFlags_NoTitleBar) ? 0 : FontSize() + GImGui.Style.FramePadding.y * 2.0f; }
-    ImGuiAabb	TitleBarAabb() const					{ return ImGuiAabb(Pos, Pos + ImVec2(SizeFull.x, TitleBarHeight())); }
-    ImVec2		WindowPadding() const					{ return ((Flags & ImGuiWindowFlags_ChildWindow) && !(Flags & ImGuiWindowFlags_ShowBorders)) ? ImVec2(1,1) : GImGui.Style.WindowPadding; }
-    ImU32		Color(ImGuiCol idx, float a=1.f) const	{ ImVec4 c = GImGui.Style.Colors[idx]; c.w *= a; return ImConvertColorFloat4ToU32(c); }
-};
-
-static ImGuiWindow*	GetCurrentWindow()
-{
-    GImGui.CurrentWindow->Accessed = true;
-    return GImGui.CurrentWindow;
-}
-
-static void RegisterAliveId(const ImGuiID& id)
-{
-    if (GImGui.ActiveId == id)
-        GImGui.ActiveIdIsAlive = true;
-}
-
-//-----------------------------------------------------------------------------
-// User facing structures
-//-----------------------------------------------------------------------------
-
-ImGuiStyle::ImGuiStyle()
-{
-    WindowPadding			= ImVec2(8,8);		// Padding within a window
-    WindowMinSize			= ImVec2(48,48);	// Minimum window size
-    FramePadding			= ImVec2(5,4);		// Padding within a framed rectangle (used by most widgets)
-    ItemSpacing				= ImVec2(10,5);		// Horizontal and vertical spacing between widgets
-    ItemInnerSpacing		= ImVec2(5,5);		// Horizontal and vertical spacing between within elements of a composed widget (e.g. a slider and its label)
-    TouchExtraPadding		= ImVec2(0,0);		// Expand bounding box for touch-based system where touch position is not accurate enough (unnecessary for mouse inputs). Unfortunately we don't sort widgets so priority on overlap will always be given to the first widget running. So dont grow this too much!
-    AutoFitPadding			= ImVec2(8,8);		// Extra space after auto-fit (double-clicking on resize grip)
-    WindowFillAlphaDefault	= 0.70f;
-    WindowRounding			= 10.0f;
-    TreeNodeSpacing			= 22.0f;
-    ColumnsMinSpacing		= 6.0f;				// Minimum space between two columns
-    ScrollBarWidth			= 16.0f;
-
-    Colors[ImGuiCol_Text]					= ImVec4(0.90f, 0.90f, 0.90f, 1.00f);
-    Colors[ImGuiCol_WindowBg]				= ImVec4(0.00f, 0.00f, 0.00f, 1.00f);
-    Colors[ImGuiCol_Border]					= ImVec4(1.00f, 1.00f, 1.00f, 1.00f);
-    Colors[ImGuiCol_BorderShadow]			= ImVec4(0.00f, 0.00f, 0.00f, 0.60f);
-    Colors[ImGuiCol_FrameBg]				= ImVec4(0.80f, 0.80f, 0.80f, 0.30f);	// Background of checkbox, radio button, plot, slider, text input
-    Colors[ImGuiCol_TitleBg]				= ImVec4(0.50f, 0.50f, 1.00f, 0.45f);
-    Colors[ImGuiCol_TitleBgCollapsed]		= ImVec4(0.40f, 0.40f, 0.80f, 0.20f);
-    Colors[ImGuiCol_ScrollbarBg]			= ImVec4(0.40f, 0.40f, 0.80f, 0.15f);
-    Colors[ImGuiCol_ScrollbarGrab]			= ImVec4(0.40f, 0.40f, 0.80f, 0.30f);
-    Colors[ImGuiCol_ScrollbarGrabHovered]	= ImVec4(0.40f, 0.40f, 0.80f, 0.40f);
-    Colors[ImGuiCol_ScrollbarGrabActive]	= ImVec4(0.80f, 0.50f, 0.50f, 0.40f);
-    Colors[ImGuiCol_ComboBg]				= ImVec4(0.20f, 0.20f, 0.20f, 0.99f);
-    Colors[ImGuiCol_CheckActive]			= ImVec4(0.90f, 0.90f, 0.90f, 0.50f);
-    Colors[ImGuiCol_SliderGrab]				= ImVec4(1.00f, 1.00f, 1.00f, 0.30f);
-    Colors[ImGuiCol_SliderGrabActive]		= ImVec4(0.80f, 0.50f, 0.50f, 1.00f);
-    Colors[ImGuiCol_Button]					= ImVec4(0.67f, 0.40f, 0.40f, 0.60f);
-    Colors[ImGuiCol_ButtonHovered]			= ImVec4(0.60f, 0.40f, 0.40f, 1.00f);
-    Colors[ImGuiCol_ButtonActive]			= ImVec4(0.80f, 0.50f, 0.50f, 1.00f);
-    Colors[ImGuiCol_Header]					= ImVec4(0.40f, 0.40f, 0.90f, 0.45f);
-    Colors[ImGuiCol_HeaderHovered]			= ImVec4(0.45f, 0.45f, 0.90f, 0.80f);
-    Colors[ImGuiCol_HeaderActive]			= ImVec4(0.60f, 0.60f, 0.80f, 1.00f);
-    Colors[ImGuiCol_Column]					= ImVec4(1.00f, 1.00f, 1.00f, 1.00f);
-    Colors[ImGuiCol_ColumnHovered]			= ImVec4(0.60f, 0.40f, 0.40f, 1.00f);
-    Colors[ImGuiCol_ColumnActive]			= ImVec4(0.80f, 0.50f, 0.50f, 1.00f);
-    Colors[ImGuiCol_ResizeGrip]				= ImVec4(1.00f, 1.00f, 1.00f, 0.30f);
-    Colors[ImGuiCol_ResizeGripHovered]		= ImVec4(1.00f, 1.00f, 1.00f, 0.60f);
-    Colors[ImGuiCol_ResizeGripActive]		= ImVec4(1.00f, 1.00f, 1.00f, 0.90f);
-    Colors[ImGuiCol_CloseButton]			= ImVec4(0.50f, 0.50f, 0.90f, 0.50f);
-    Colors[ImGuiCol_CloseButtonHovered]		= ImVec4(0.70f, 0.70f, 0.90f, 0.60f);
-    Colors[ImGuiCol_CloseButtonActive]		= ImVec4(0.70f, 0.70f, 0.70f, 1.00f);
-    Colors[ImGuiCol_PlotLines]				= ImVec4(1.00f, 1.00f, 1.00f, 1.00f);
-    Colors[ImGuiCol_PlotLinesHovered]		= ImVec4(0.90f, 0.70f, 0.00f, 1.00f);
-    Colors[ImGuiCol_PlotHistogram]			= ImVec4(0.90f, 0.70f, 0.00f, 1.00f);
-    Colors[ImGuiCol_PlotHistogramHovered]	= ImVec4(1.00f, 0.60f, 0.00f, 1.00f);
-    Colors[ImGuiCol_TextSelectedBg]			= ImVec4(0.00f, 0.00f, 1.00f, 0.35f);
-    Colors[ImGuiCol_TooltipBg]				= ImVec4(0.05f, 0.05f, 0.10f, 0.90f);
-}
-
-ImGuiIO::ImGuiIO()
-{
-    memset(this, 0, sizeof(*this));
-    DeltaTime = 1.0f/60.0f;
-    IniSavingRate = 5.0f;
-    IniFilename = "imgui.ini";
-    LogFilename = "imgui_log.txt";
-    Font = NULL;
-    FontAllowScaling = false;
-    MousePos = ImVec2(-1,-1);
-    MousePosPrev = ImVec2(-1,-1);
-    MouseDoubleClickTime = 0.30f;
-    MouseDoubleClickMaxDist = 6.0f;
-}
-
-// Pass in translated ASCII characters for text input.
-// - with glfw you can get those from the callback set in glfwSetCharCallback()
-// - on Windows you can get those using ToAscii+keyboard state, or via the VM_CHAR message
-void ImGuiIO::AddInputCharacter(char c)
-{
-    const int n = strlen(InputCharacters);
-    if (n < sizeof(InputCharacters) / sizeof(InputCharacters[0]))
-    {
-        InputCharacters[n] = c;
-        InputCharacters[n+1] = 0;
-    }
-}
 
 //-----------------------------------------------------------------------------
